NF: move UpdateValuesFromDeck where its used
